Add reactive semantic browser players
